3rdparty/t9write: Remove unused code
t9writeworker.cpp contained old unused code for result
processing. Remove it.

The instant gesture detection is useless in worker thread,
since it is meant to be used in main thread for a specific
use case in SCR mode. We don't need it, so remove it.

Finally, the abort functionality of the recognition call
was tested by calling the recognize function with interrupt
function. This will not work in v8.0.0 anymore, where the
recognize function will return error for all subsequent calls,
even by setting the interrupt function to NULL. Instead, we
now use a proper compile-time functional check.

Change-Id: Ic98fd3b45912f29a4c72e3650f98b7b520c4ade9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>