3rdparty/t9write: Fix non-blocking recognition
Previously the handwriting recognition with T9 Write was blocking
at the beginning of new handwriting arc. This was due to the fact
that while the recognition was done in background task, the arc
addition was done in main thread, so we'd had to wait for the
background recognition task to complete to avoid accessing the
T9 Write session in parallel.

This change moves the arc addition to background task, so the main
thread does not have to block for the background tasks while the
handwriting is ongoing.

The rationale for this change is the new UCR input mode, which
consumes a lot more CPU during recognition, and the UI will become
non responding.

[ChangeLog] Optimize UI performance by eliminating the need to wait
for previous results to be completed during T9 Write handwriting.

Change-Id: I8d0cecfa4599b63ccb21c1f6a691e5f21c93158c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>