3rdparty/t9write: Don't discard pre-edit text on manual key input
When entering a manual key (e.g. period) on T9 Write layout,
the current pre-edit text was discarded. The correct operation
is to commit the current pre-edit text.

This was not noticed until now, since the punctuation works so
well with T9 Write.

[ChangeLog] Fixed an issue in the T9 Write input mode that caused the
current word suggestion to be rejected when a character was entered
through the keyboard.

Change-Id: I1fb9e5207a47ef2f239e6a3ba9b2acfb269fa63b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>