Fix build of pinyin plugin on platforms without pthread
...such as VxWorks. Normally, the Qt dependency pulls in pthread,
because the qt feature adds the thread feature (which in turn adds
pthread and other necessary flags to the compiler and linker command
lines), but in this case we don't have Qt so it must be manually added.

Use the higher level abstraction instead of the direct -pthread flag,
so that the compiler and linker flags used are taken from mkspecs.

Task-number: QTBUG-63516
Change-Id: I88876810129078dfe4898314e45379ff87f95b80
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>