tst_inputpanel: don't rely on index var being valid after first loop
In Qt 5.10 this code would work, but in dev inputIndex is null.
It might be a regression or it might be a fix to comply with the
standard, but either way, we should be explicit in the test and
initialize the variable in the for loop's initial expression.

Task-number: QTBUG-64988
Change-Id: Idf6019307a02a32dcb6237e2e83e5a9cd512df3d
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Liang Qi <liang.qi@qt.io>