Allow InputMethod to process pre-edit click
We already have InputMethod::reselect, but it is
limited to case where the current pre-edit is
discarded and new range is made active.

However, if an input method wants to move the
cursor within pre-edit, it is not possible with
the current implementation.

This change adds the new API for the InputMethod
which allows to handle pre-edit click before
reselect happens. If the input method handles
the event, reselect will not be done.

This change also improves the cursor handling
within pre-edit text. Previously if the cursor
attribute was added to pre-edit text and the
input was committed, the cursor would jump to
position after pre-edit text. This is now taken
into account when committing text and cursor
location will be preserved.

Change-Id: I0a32f75eb4e454c86dd8d4f4016ac02fc1bc6c60
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>