basic example: ensure AutoScroller scrolls correctly when first opened
In AutoScroller.qml, the following line

    target: inputItem && !Qt.inputMethod.animating ? Qt.inputMethod : null

prevents onKeyboardRectangleChanged from getting executed. Even though
Qt.inputMethod.animating is updated before the rectangle is updated,
the handler still doesn't get emitted. Since the VKB only emits
keyboardRectangleChanged when the input method is not animating (see
the binding in Keyboard.qml), we can safely remove that check.

That just leaves one more issue: clicking in an input field, then
clicking outside of it (on the MouseArea) to hide the keyboard, then
clicking back in a lower text field will not cause AutoScroller
to scroll. In this case, onInputItemChanged is called, but the if
checks (mentioned at the top of the description of the bug report) use
the full height of the flickable before it's reduced to account for the
keyboard. Subtracting the height of the keyboardRectangle from the
flickable's height didn't work, so that issue needs to be fixed
separately.

Task-number: QTBUG-66294
Change-Id: I3733bb6be05443748eb873d76daee428be9687e8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>