FullScreenMode: Send return key without modifiers
When the return key is sent to a multiple line text field it can be
interpreted differently if the Shift modifier is included. This makes
sense for a hard keyboard as you can use it to enter in a new line
versus a new paragraph. For the virtual keyboard however, it is best to
just send it as a plain return so it handles it correctly.

Change-Id: I995504d7ab6961fd8baf016d2d68be4b659d1e19
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>