plugins/t9write: Fix operator precedence
Fix compiler warning:

operator '?:' has lower precedence than '+'; '+' will be evaluated first

This bug caused the function to return either 1 or 0 when the expected
return value was the number of tasks. However, the bug did not affect the
functionality since the return value was always compared to zero. Now
the function returns the expected result.

Change-Id: Id3c68029542c6494928ceb313ea5d41a3676829b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>