Add user dictionary and learning for Hunspell
This change adds user dictionary and learning function for Hunspell.

Learning happens when the user selects the first candidate from the
word candidate list (or presses the space key while the first candidate
is selected) and the word does not exist in the default dictionary.

User can remove a word from user dictionary by long pressing an item
on the word candidate list (and selecting from the menu). This also
allows the user to block words originating from the system dictionary.

The Hunspell user dictionary is hard limited to 100 words. However, when
user enters a word again (which exists in the user dictionary), it will
be prioritized in the dictionary, so the most frequent words stay in the
dictionary.

The dictionaries are language and locale specific and are stored in:

    QStandardPaths::GenericConfigLocation + "/qtvirtualkeyboard/hunspell"

The dictionaries are UTF-8 encoded text files and contain one word per line.

The same directory also contains the blacklist files (which also exist per
language and locale). These files contain words which are blacklisted from
the default dictionary.

[ChangeLog] Added user dictionary and learning for Hunspell

Change-Id: Ib0fc70f7eaa14ec49b74000144a75c59313ac0fb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>