Add layout fallback feature and eliminate duplicate layouts
Keyboard layouts such as symbols and handwriting have been duplicated
commonly. While it has been possible to avoid duplication with digits,
numbers and dialpad layouts, it was not possible with handwriting
layout, because the existence of handwriting layout indicates the
language supports handwriting (which is used to enable the handwriting
button on the main layout).

This change modifies the fallback mechanism so that a special fallback
file called <layout type>.fallback must exist in place of the layout.
This allows to explicitly state whether a layout fallback should be
used.

This change also removes the duplicated layouts.

Change-Id: I22764f96a992c92731128fa192365dec419d17eb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>