Trim InputContext API
The purpose of this change is to reduce the number of public API in the
virtual keyboard.

This change moves a lot of stuff from InputContext to ShiftHandler and
InputContextPrivate and exposes the private API to QML through
InputContext.priv property.

Almost all the unrelevant APIs were moved away, except some properties
and methods needed by selection control. These were left intact because
moving them is not trivial.

Change-Id: I1f23f5f54bc21c68996cb220a66d16d34b5d14ce
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>