Fix regression in test_wclAutoCommitWordSetting
Commit e803aec1ea21fd00e13b9535a4b536cc43c26ee4 modified
the test case and caused regression. The idea was correct, but
the implementation used SignalSpy for model.dataChanged, which is
ambiguous, since there are multiple signals per update.

This change introduce inputMethodSelectionListChangedSpy which
monitors selectionListChanged signal from the input method. This
is reliable, as long as suggestions are generated background.

Change-Id: Ica2710b8c389c13e1a1b1d28027667e031bc4f59
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>