Fix handling of invalid VirtualKeyboardSettings.activeLocales
If VirtualKeyboardSettings.activeLocales contains value(s) but
not any value which is actually available, the available locale
was set to contain only the first installed locale. Previously,
before the layout fallback improvement, it was set to "en_GB"
only.

This change fixes the issue and treats invalid activeLocales
as if it is empty (default). This means the default locale is
set as normal.

Change-Id: I8377d0f28d9217b896aa446fe6037f32943e72f4
Reviewed-by: Andy Shaw <andy.shaw@qt.io>