Fix resource initialization in static build
Resources used by qtvirtualkeyboard must (still) be initialized manually
in static build. The reason for this is that the resources are located
in the virtualkeyboard module, not in the plugin.

Layouts resources in plugins must have unique name. This change renames the
layouts resource file in each plugin. Otherwise the Q_INIT_RESOURCE does not
work, since all use the same name.

Finally, rename the static resource files to avoid using too generic name and
clash with the user application.

Change-Id: I53054605f14ff8a7b61ce2736f61e456b6c3ee88
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>