Do show module license on all help pages
Qt Virtual Keyboard has a different main license than most other Qt
modules, which is easy to miss. We do show the license on the module
landing page and in the 'All modules' overview page, but if you
directly open an example or a class page you can still miss this.

This change adds the main license description to the header shared
by all pages of the module; we (mis)use the buildversion variable for
that, which was so far unused.

Right now qdoc does not 'linkify' the entry. It could add a generic
link to the landing page though, where the license is also described.

Task-number: QTBUG-70354
Change-Id: I5f271b1bcba6992abfd1ad7e213fc5f2a93a0d25
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>