Automatically register the latest import version
This follows what was done in d28c9f6a for Qt Quick Controls 2.
The latest import version (e.g. 2.14 in Qt 5.14) will automatically be
registered whenever the Qt version is bumped.

This avoids needing to wait until a new type is added (or a new
revision is added to an existing type) before being able to use the
newest Qt Quick version.

Change-Id: Ic2b59f50bfae8fed5148eae9596578adb0aadb52
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>