Qt virtual keyboard wrapper for X11
This application is a Qt virtual keyboard wrapper that uses Qt virtual
keyboard for any X11 input. Application uses At-Spi2 library for
listening focus-in events on text widgets and pops up the Qt virtual
keyboard when an accepted focus-in event is received. There's also
support for listening focus-in events from Chromium browser. For
mapping and sending the key events to focused text widget the application
uses libXtst, libX11 and libxdo which are also
external dependencies to the application among At-Spi2 library.

Change-Id: Ifa54969ab4f61e1d48f11ca11058c836f2a403b8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>