Remove usages of deprecated qBinaryFind
Note, that qBinaryFind is replaced by std::lower_bound, rather than
std::binary_search, since the later returns only a bool indicating if
an item was found. But in this case we need to access the found item,
so std::lower_bound (which does the same thing) is used instead.

Task-number: QTBUG-76491
Change-Id: I1ccc3c82738c8dce91df414f442bd011469462c7
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>