Port from implicit QChar(*int*) to explicit ones
Found a nice misuse of QChar <=> Qt::Key, which happens to work for
'\'', but not for all other keys (e.g. Esc)...

I want to note that, yes, I took note of the code's abuse of Qt
containers, and no, I _did_ constrain myself; this patch is just
fixing implicit *int*->QChar conversions.

Change-Id: I3cb43417f3922fdc500783b7e0b8caecc0d3c5f5
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>