Fix lipitoolkit build on Ubuntu
-L$$OUT_PWD/../../../../lib points to:

/home/mitch/dev/qt5.13-debug/qtvirtualkeyboard/src/plugins/lipi-toolkit/plugin/../../../../lib

which expands to:

/home/mitch/dev/qt5.13-debug/qtvirtualkeyboard/lib

I have no lib folder there.

This patch changes the pro file to use QT_HOST_LIBS
(/home/mitch/dev/qt5.13-debug/qtbase/lib/) and not OUT_PWD, as OUT_PWD
refers to where the Makefile is being created in and not where the
libraries are actually being built into. Since the libraries are
placed where the module/lib path is, using QT_HOST_LIBS will
ensure it is in the same location.

Change-Id: I1921b18d9362c49174c5ba6b267fe194a963d91c
Reviewed-by: Andy Shaw <andy.shaw@qt.io>