Allow shift toggle with Qt::ImhNoAutoUppercase
Qt::ImhNoAutoUppercase means that the input method should not try to
automatically switch to upper case when a sentence ends. QtVKB was
using it also to make shift work as caps lock. Change the behavior
so that the shift key affects only the next character.

[ChangeLog] Using Qt::ImhNoAutoUppercase hint (used also with password
fields) no longer makes shift key function as a caps lock.

Change-Id: I231c8d192a67e19f04bc4d3fcf0bc9c676a12830
Task-number: QTBUG-80058
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>