Make sure that the input panel receives events during modal sessions
Parent the input panel to the focus items' window's contentItem, with
highest z order. This makes the input panel a sibling of the popup's
overlay and dimmer items, but thanks to the highest z-order, the input
panel gets all events.

This requires that we store the input panel in the InputContext, which
again assumes that there is only one input panel. Making this assumption
explicit with a Q_ASSERT.

Also, since __isRootItem is now true, we need to override the reliance
on this property alone to avoid that SelectionControl becomes disabled.
Introduce a second internal property. This should be followed up by a
change that makes the disabling of the SelectionControl in the desktop
integration case an explicit step.

Fixes: QTBUG-56918
Done-with: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Change-Id: I58bbbb0adf145c9cfc30a7dd0acb2a0e1231481b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>