Refactor desktop-detection logic to become an explicit property
The __isRootPanel internal property was used to detect whether the
keyboard is running in as a stand-alone, toplevel window in the desktop.
The value of the property depended on the location of the keyboard panel
within the QQuickItem hierarchy, which is fragile.

Instead, declare an explicit property desktopPanel that is set to true
by the desktop integration when the panel has been created.

Change-Id: Idb9e90e01e9b1c643e84287716e0b964256b115e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>