Auto-capitalize only after space following a sentence-ending character
A dot alone should not change to capital letters, only the space
following the dot. This is standard behavior on mobile platforms.

In inputmethods that support auto-completion and auto-space-insertion,
we need to watch out if the auto-inserted space triggered auto-
capitalization. If so, then the inserted text needs to be
capitalized even though the keyboard will have shown lower-case
characters.

Fixes: QTBUG-77673
Change-Id: Icd907055d6557a7756468318fba5669eb8f62a28
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>