Initial CMake port of QtVirtualKeyboard
This port makes sure that all the features that are currently tested in
the CI are complete and that the tests pass.

lipi-toolkit can't be tested since it's currently broken in qmake.

Myscript wasn't tested as it is also not tested in coin and the SDK is
not available.

T9Write wasn't tested as the SDK is not avaialble.

Removed outdated src/virtualkeyboard/styles/styles.pro.
The styles seem to be in src/styles now.

Change-Id: Ibba41d592c51a19dee63840c02a945ec7b1d721e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>