Recycle drawing canvases to save resources v3
Each handwriting trace requires a canvas. Previously the canvases
were destroyed after the trace is discarded. This might not be the
optimal way to handle the canvas component.

This change modifies the canvas creation and preserves the existing
canvases and recycles them for reuse. Also, a handful of canvases
are now created at startup to reduce the latency at initial drawing.

The total number of canvases is now limited as well. If the inventory
runs out and the maximum number is reached, the oldest drawing will
be reused, even if it was visible on the screen.

Due to fact that the Trace object may be unreferenced before they are
discarded, it is now mandatory to set ownership to QVirtualKeyboardTrace
objects.

v2: Fix the order in which the canvas recycling happens
v3: Fix bug where the canvases were not actually recycled

Change-Id: Ia26bc96ab1405a578a29420e419abe8e420fb712
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>