Fix position of the keyboardRectangle in app integration
InputPanel contained double mapping of keyboardRectangle to QML
root. The problem was that inputPanel.x and y were already relative
to parent so mapToItem(null) doubled the value.

Use function to calculate the rectangle. This function also
references the position of the inputPanel, so it is properly
updated in the Binding.

Task-number: QTBUG-85554
Pick-to: 5.15
Change-Id: Iaf5075f31f27bc08bbf41aa5960303705135cd16
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>