Replace layoutChangeNotifyTimer with Qt.callLater
The layoutChangeNotifyTimer is used internally by the Keyboard to delay
the notification of keyboard changed event to keyboard observer
component. This is to avoid subsequent notification calls.

Use of timer is not perhaps the best way to accomplish this. This
change replaces the timer with Qt.callLater, which will do the same but
without the additional delay.

Pick-to: 6.1
Change-Id: I9bbefb47ddd0daa64096bd676c25da4e94bcd33d
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>