Decrease the z-value of the InputPanel to 10000
There are valid use cases where UI elements need to be on top of the
InputPanel, so occupying the maximum Z value is unnecessary.
A Z value of 10000 should be a reasonably default and leaves enough
room on both sides. Document the new default.

[ChangeLog][InputPanel] The default z-value of the
InputPanel item has been decreased to 10000.

Fixes: QTBUG-92881
Change-Id: Ifab12273746e5a33e14d8346e03af6ec80c2b049
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>