Make sure keyboard layout observer returns correct geometry
The scanLayout() function may be called at any time, even during the
layout process. This means, that some parts of the layout may overlap
the layout geometry.

The user (e.g. the XT9 layout engine) reports an error if any of the
keys overlap the layout geometry.

This change adjusts the geometry of the reported layout if necessary.

Change-Id: I8d2ccec695484518c650a29da33e02de5e513126
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>