plugins/cerence-hwr: Fix waiting of recognition results during key event
Key events (from virtual keys) are captured and handled by
T9WriteInputMethod. If the recognition results are currently being
processed, it will block and wait until results are handled.

Unfortunately, the wait function was broken, because it actually failed
to process the results.

The results were processed by a slot connected to a signal via queued
connection, so it was impossible to wait without breaking out from the
event loop.

Fix the issue by separating the resultList handover to a slot connected
via direct connection and the resultList processing to a slot connected
via queued connection. This makes it possible to wait for resultList
handover process the results immediately.

Pick-to: 6.2
Change-Id: Ia7be4ddcc5071775d43a7dddf36e7c33cf1d1e96
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>