Only reparent the InputPanel when necessary
The input panel needs to work also while a modal popup is open. Since
the concept of modality does not exist in Qt Quick, and is implemented
with an event-eating overlay, 4e8b3dd45ae4cc66a1b77cce901f80406b2a0f69
introduced the automatic reparenting of the integrated input panel so
that it a sibling of the overlay, with a higher z-value.

This effectively changed the parent item, and had negative side effects
to mechanisms such as anchoring.

This change only reparents the input panel if there is an overlay (using
the mechanism also used by the Qt Quick Control class QQuickOverlay)
that is visible when the input item changes, and sets the z-value to be
one above the overlay's. The prevents modal blocking during the overlay
session. When the input item changes again and the overlay is no longer
visible, then the parent and z value get reset.

This improves the situation, but is still not ideal, as during modal
sessions the anchoring of the input panel will still not work as
expected.

Add a test case that verifies that clicks go through to the keyboard
while a modal dialog is open.

Fixes: QTBUG-97075
Task-number: QTBUG-56918
Task-number: QTBUG-92881
Change-Id: I7a5bcfebe7b69780234737a2311e08b2a6f60a4e
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
(cherry picked from commit 5564f1c96322d7c3f884d2e6f9b3a95e3134fb9f)
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>