Update documentation wrt modality and z ordering
As of 1129af130964c5e0f2db6e8a6fb6dc2fe70f3bc8, the InputPanel uses a
containment masks on the overlay popup to keep it enabled even during
modal sessions. It no longer reparents itself to the overlay item and
doesn't change the z value, so remove that documentation.

Add instead that the one, common input panel is bypassing modality,
and that it may be obscured by items with a higher z value.

Task-number: QTBUG-97439
Pick-to: 6.2 6.3
Change-Id: I493d5446186dae48523f63b29a473ee83daa4ffb
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>