Use proper type (instead of var) for Keyboard.activeKey
Virtual Keyboard was not impacted by this bug until it started using
qmlcachegen with combined qmlmodule.

The bug triggered with following property and code:

    property var activeKey

    ...

    activeKey = functionThatReturnsAnItem()

    ...

    if (activeKey) {
        // Unreachable code
    }

The reason is that qmlcachegen incorrectly handled implicit type cast.

This issue does not impact on any of the previous released qt versions,
since none of them were using qmlcachegen for virtual keyboard
components. This is because the components were embedded in the
resources of the virtual keyboard qt module (QtXVirtualKeyboard).

The preferred workaround in this case is to use an explicit type
instead of variant. This will eliminate the implicit type cast.

Task-number: QTBUG-100883
Change-Id: I4fea27df2d1242f6bdb38a4f4fddd44a8d5902fc
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>