Update commercial license headers
Updated header.COMM to the files in tqtc-qtvirtualkeyboard.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4937
Change-Id: I4ada1c1fc8fee61ce0855acc72a5489f6851788a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>