Fix screen orientation handling for basic-b2qt example
Since Qt 6.2.2 the InputPanel is not reparented anymore for modal
dialogs. The basic example contains code to handle orientation for this
case. Now, it is enough to handle orientation at the application window
level. The InputPanel orientation will follow this automatically.

In addition, for QTBUG-106895, the example uses
Screen.orientation instead Screen.primaryOrientation to detect
landscape orientation. It seems on Mac the Screen.orientation can be
Qt.PrimaryOrientation, which was incorrectly handled as portrait
orientation.

Fixes: QTBUG-106895
Task-number: QTBUG-83217
Pick-to: 6.4.0 6.4 6.2
Change-Id: Ic40d9194150334f31a236f0b6af84d59f914759c
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>