Fix pre-edit text on shadow input when reselecting the first word
The shadow input control did not receive pre-edit text when the first
word was reselected. In this case, the text is transferred from the
input field to pre-edit field.

Analysis shows that QInputMethodEvent::Selection cannot be used
together when complete text buffer is replaced. This is because of
quirks found in QQuickTextInput's processInputMethodEvent.

Fix by handling text replacement and selection as separate events.

Change-Id: I2f730a491685e55ff28a98a4929a1cd1b5bf0e6a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Inho Lee <inho.lee@qt.io>
(cherry picked from commit 960304256ece1ab019eeab91697317b1abf303ae)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>