Improve shadow input control
Previously, the shadow input control was taking up the remaining space
above the keyboard, resulting in an unsightly appearance. Furthermore,
this layout was unnecessary.

With this update, the shadow input has been redesigned to display as a
single line edit above the keyboard, expanding as needed. If the input
exceeds the width of the keyboard, it becomes flickable.

Overall, this redesign results in a cleaner and more user-friendly
interface, improving the user experience.

Fixes: QTBUG-60898
Fixes: QTBUG-70675
Change-Id: I58a64d2f3ceb057ebdf99205dc4622181a37d784
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>