Fix full screen input control sync (disable when not active)
The full screen input (aka shadow input) was continued being
synchronized even after the full screen mode was disabled. Not only
this caused an unnecessary overhead, it resulted in incorrect behavior
in some cases when initializing cursorPosition and selection.

This caused some of the test to fail without obvious reason. The root
cause was the cursor synchronization of the shadow input setting the
cursor during the tests. Now the shadow input item is reset when full
screen mode is disabled. In addition, the cursor synchronization is
not run when full screen mode is toggled.

Remove affected test cases from the blacklist.

Fixes: QTBUG-97901
Task-number: QTBUG-97830
Pick-to: 6.5 6.4
Change-Id: I571433adfdb5a3f3cd9d994f0d6e916cae372054
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Inho Lee <inho.lee@qt.io>