Fix collapsed keyboard layout caused by zero width input panel
This patch fixes the problem by delaying the loading of the layout
until the keyboard width is valid.

An additional updateInputMethod call is necessary, as the input method
is dependent on the keyboard layout. However, this should only be
called if the inputMethod is not set. This can occur with
DesktopInputPanel.

Furthermore, the restoreMode has been removed from the keyboard layout
binding, as the keyboard layout should not be reset.

Pick-to: 6.5
Fixes: QTBUG-111928
Change-Id: I94aae3dc978e3c0445edd2d55fcae233a43d581a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Inho Lee <inho.lee@qt.io>