Fix InputPanel keyboardRectangle binding
The word candidate view was previously accounted to InputPanel.Keyboard
height based on visibility. This caused complex binding issues during
animations. The word candidate view was also included into
keyboardRectangle. This prevented the application from comparing it
to the cursorRectangle in binding fullScreenMode on the result.

With this change the word candidate list is not accounted to keyboard
height anymore. Instead, the height is added to the InputPanel.height.
Furthermore, the keyboardRectangle never includes the word candidate
list. This allows the application to compare it to the cursorRectangle.

Fixes: QTBUG-97830
Change-Id: I40de1bbc9a79da3e0a0cc1de70836bc8bd3289ff
Reviewed-by: Inho Lee <inho.lee@qt.io>
(cherry picked from commit 2bd67e9248d0a11d16fe4eec767e663c8e47939e)