Ensure that the language popup is within keyboard bounds
Currently the language popup is always located correctly, because it is
opened from a keyboard key located away from the edges.

However, if the key is at the edge, the language popup clips with
keyboard boundaries.

This change adjusts the offset so that the popup is within keyboard
area.

Task-number: QTBUG-113024
Change-Id: I9c58658130e1c8806827b765c8be4ba364962318
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>