Doc: Add a DocBook version of custom svg images
QDoc provides a "\raw" command that allows a user to inject
format-specific code that complements the code that QDoc generates.

The "\raw" command expect the user to provide a "format" argument to
ensure that the provided code is generated only for the format that is
written for.

Due to a bug in QDoc DocBook generation, the DocBook documentation was
generating, sometimes incorrectly, "\raw" commands intended for the HTML
format.

The Qt virtual keyboard documentation uses the "\raw HTML" commands to
provide some custom SVGs.

Due to the aforementioned bug, QDoc was generating some equivalent code
for DocBook, albeit no user-provided DocBook equivalent existed.

As the bug is now being fixed, the DocBook documentation for virtual
keyboard will lose the custom SVGs in its output.

To preserve the SVGs for when generating DocBook documentation, an
equivalent "\raw DocBook" command is now provided.

Change-Id: I272163cd88b133e07fb790293baa9d4833429087
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>