Make qvirtualkeyboardnamespace.h private
This change was made based on feedback on Qt 6.6 api review.

Since the qvirtualkeyboardnamespace.h is currently used as private API,
it makes no sense to keep it public.

[ChangeLog][Potentially Source-Incompatible Changes][VirtualKeyboard]
The qvirtualkeyboardnamespace.h header has been made private. None of
the enum types declared in that header are used in public C++ APIs.

Pick-to: 6.6
Change-Id: I889bf71bb36d00140dcffad7e672452c93eaaa98
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>