Fix binding loop warning emitted by inputpanel test
zh_TW has only two supported input modes Zhuyin and Cangjie.

At the time locale is set to zh_TW and the main layout is loaded,
inputMode is not yet set to Zhuyin or Cangjie, because the input
method is created by the layout and the default input mode is
defined by the input method.

The binding loop happens, because in the test case the inputMode
is Latin during creation:

    create: inputMode === Latin -> Cangjie
    default: inputMode === Zhuyin -> Zhuyin
    test: inputMode === Cangjie -> Cangjie

Fix by setting sourceComponent to null if the inputMode is not yet
set to supported input mode.

Pick-to: 6.5 6.6
Change-Id: I1c9bcccc5655e3ec1553e6e5414b47733eba5b82
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>