Update the basic example to deploy the required plugins explicitly
Add the missing Qt6::QAndroidMediaPlugin plugin explicitly to the
basic example using the qt_import_plugins call. This is used when
deploying plugins on Android platform.

TODO: QtMultimedia module is not used by the application directly, but
it's required to link it, to make the qt_import_plugins cal for
QAndroidMeadiPlugin work. See QTBUG-122607.

Fixes: QTBUG-121658
Pick-to: 6.7 6.6 6.5
Change-Id: I5c45541e5ceed556d5d995b1cb4ff1372859a16b
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>