Move test setup to setup classes
Some tests in qtvirtualkeyboard depend on the QT_IM_MODULE env variable.
The tests use static initialization of variables to set it before main is called.
Unfortunately, the call to setenv before main starts does not work as intended
on VxWorks. The function is called, but the test does not see that the var has
been set.

Fix the issue by moving the init step to a setup class.

Task-number: QTBUG-115777
Pick-to: 6.7
Change-Id: I4651928dfceae7e5e3d1a98ea1da2e8c6cc5f217
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>