Skip test_blockHover on offscreen platform; stop hovering Button
Skip it as in qtdeclarative 653ad25d64b06686e76391ff35968c1a652941d2
and others: it's a known issue that there is no hover on offscreen.

It's uncertain why mouseMove(10, 10) would fail to hover the Button
on some platforms, but what we really need to test is that it's
_not_ hovered when moving the mouse over the InputPanel on top.
So removing the sanity check may stabilize the test.

Pick-to: 6.8
Change-Id: I6a10a541cc869a196f38c41866166e4ca4144ab9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 9d9ae922a94a365ec4c8997193dfc3e49dc5a93b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>