examples: Remove AutoScroller and use the shadow input instead
AutoScroller is behaving quite bad when the keyboard is active and the
user scrolls the window. The scrolling jumps irradically.

With the new inline shadow input control (a.k.a. fullScreenMode), the
AutoScroller becomes redundant, as the contents of the obscured edit
field can be shown with the shadow input control.

Task-number: QTBUG-60898
Task-number: QTBUG-70675
Change-Id: I000a3080c7c3422342c1a81136df5bb6027f1c68
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>