Fix fullScreenMode on desktop and improve binding for keyboardRectangle
Update keyboardRectangle to include the shadow input control when in
full screen mode. This fixes the invisible shadow input on the desktop
integration.

Account for inputPanel.x and inputPanel.y in app-integrated input
panel. The keyboardRectangle is now updated when the input panel is
moved (or window resized) on the application integration.

Reduce keyboardRectangle updates when the keyboard is hidden with the
word candidate view visible. Removed "keyboard.active" condition for
wordCandidateView.visibleCondition. This reduces the number of updates
for the keyboardRectangle when the keyboard is hidden while the word
candidate view is visible.

Update test cases for keyboardRectangle to cover the word candidate
view.

Fixes: QTBUG-126402
Fixes: QTBUG-137731
Pick-to: 6.10 6.9.3 6.8 6.5
Change-Id: Iab84f71559185b8abe5af98b04ee364a6e6da193
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>