pinyin: Fix crash with "nss" input
qsort is not stable across platforms when comparing equal strings.

On Windows, the sort reordered items with identical strings, while on
Linux the original order was preserved. The intended order actually
depends on the additional field LmaPsbStrItem::lpi::psb.

Update the comparator to sort by psb in ascending order when strings
are equal.

Also add a regression test for the "nss" input.

Fixes: QTBUG-124178
Pick-to: 6.10 6.9.3 6.8 6.5
Change-Id: I12802e3f406c300c0e2564d4750f58f03fb94626
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>