Update word candidate view visibility when keyboard is hidden
Commit ab930c6d4eba91c165f2d1bb58cf48fc0555cd8f changed the behavior of
word candidate view visibility in an attempt to reduce updates to
Qt.inputMethod.keyboardRectangle. It removed the keyboard.active
dependency from visibleCondition. While this achieved the intended
goal, it introduced a regression; the input panel briefly flickers when
it is reshown after the word candidate view was hidden while the panel
was invisible.

Pick-to: 6.10 6.9.3 6.8 6.5
Change-Id: I2aeba324c0f8a0178972f7eac7263e6f511f54d7
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>